pyproject.toml
[tool.poetry]
name = "rotate-that-batch"
version = "0.1.0"
description = "When your videos are sideways, it's time to 'Rotate That Batch'."
authors = ["Kevin N. Murphy <knmurphy@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "rotate_that_batch", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
textual = "^0.38.1"
rich = "^13.5.2"
typer = "^0.9.0"
pyexiftool = "^0.5.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.1"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.1"
pytest-asyncio = "^0.21.1"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
rotate-that-batch = "rotate_that_batch.main:app"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "pass",
    "except ImportError:",
]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
strict = true
ignore_missing_imports = true

[tool.flake8]
max-line-length = 100
extend-ignore = "E203, W503"
exclude = [".git", "__pycache__", "build", "dist"]
